requirements.txt
Django==3.2
psycopg2-binary
djangorestframework==3.12.4